Moved everything to the component folder inside the component-library.
