+++ src/app/component-library/components.module.ts
@@ -1,8 +1,8 @@
 /* @format */
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ToggleComponent } from './components/components/toggle/toggle.component';
-import { ButtonComponent } from './components/components/button/button.component';
+import { ToggleComponent } from './components/toggle/toggle.component';
+import { ButtonComponent } from './components/button/button.component';
 
 @NgModule({
   declarations: [ToggleComponent, ButtonComponent],
