+++ src/app/component-library/components/toggle/toggle.stories.ts
@@ -6,7 +6,7 @@ import {
   StoryObj,
 } from '@storybook/angular';
 import { action } from '@storybook/addon-actions';
-import { ComponentsModule } from '../../../components.module';
+import { ComponentsModule } from '../../components.module';
 import { ToggleComponent } from './toggle.component';
 
 const meta: Meta<ToggleComponent> = {
